+++ src/datatypes/activity.ts
@@ -4,3 +4,4 @@ type Schemas = components["schemas"];
 
 export type Activity = Schemas["Activity"];
 export type ActivityType = Schemas["ActivityType"];
+export type ActivitiesPage = Schemas["ActivitysPage"];