+++ src/datatypes/annotation.ts
@@ -5,3 +5,4 @@ type Schemas = components["schemas"];
 export type Annotation = Schemas["Annotation"];
 export type CreateAnnotation = Schemas["CreateAnnotation"];
 export type UpdateAnnotation = Schemas["UpdateAnnotation"];
+export type AnnotationsPage = Schemas["AnnotationsPage"];