+++ src/datatypes/api-token.ts
@@ -3,8 +3,9 @@ import type { components } from "@/schema/api.js";
 type Schemas = components["schemas"];
 
 export type ApiToken = Schemas["ApiToken"];
-export type ApiTokenWithSecret = Schemas["ApiTokenWithSecret"];
+export type ApiTokenWithJWT = Schemas["ApiTokenWithJWT"];
 export type ApiTokenType = Schemas["ApiTokenType"];
 export type CreateApiToken = Schemas["CreateApiToken"];
 export type UpdateApiToken = Schemas["UpdateApiToken"];
 export type TokenExpiration = Schemas["TokenExpiration"];
+export type ApiTokensPage = Schemas["ApiTokensPage"];