+++ src/datatypes/comment.ts
@@ -5,3 +5,4 @@ type Schemas = components["schemas"];
 export type Comment = Schemas["Comment"];
 export type CreateComment = Schemas["CreateComment"];
 export type UpdateComment = Schemas["UpdateComment"];
+export type CommentsPage = Schemas["CommentsPage"];