+++ src/datatypes/pagination.ts
@@ -2,4 +2,4 @@ import type { components } from "@/schema/api.js";
 
 type Schemas = components["schemas"];
 
-export type Pagination = Schemas["Pagination"];
+export type CursorPagination = Schemas["CursorPagination"];